requirements.txt
streamlit
keras
tensorflow
numpy
pillow